+++ experimental-examples/tina-cloud-starter/.tina/schema.ts
@@ -11,7 +11,7 @@ const schema = defineSchema({
     media: {
       tina: {
         publicFolder: "public",
-        syncFolder: "uploads",
+        mediaRoot: "uploads",
       },
     },
   },